+++ src/Files.App/Actions/Content/ImageEdition/RotateLeftAction.cs
@@ -5,7 +5,6 @@ using Files.App.Contexts;
 using Files.App.Extensions;
 using Files.App.Helpers;
 using Files.App.ViewModels;
-using Files.Backend.Services.Settings;
 using System.ComponentModel;
 using System.Linq;
 using System.Threading.Tasks;
