+++ src/Files.App/Helpers/AppThemeResourcesHelper.cs
@@ -1,6 +1,6 @@
 using CommunityToolkit.WinUI.Helpers;
+using Files.Backend.Services;
 using Files.Backend.Services.Settings;
-using Files.Shared.Services;
 using System;
 
 namespace Files.App.Helpers
