+++ src/Files.Shared/Services/IJumpListService.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Files.Shared.Services;
-
-public interface IJumpListService
+namespace Files.Shared.Services
 {
-	Task AddFolderAsync(string path);
-	Task RefreshPinnedFoldersAsync();
-	Task RemoveFolderAsync(string path);
-	Task<IEnumerable<string>> GetFoldersAsync();
+	public interface IJumpListService
+	{
+		Task AddFolderAsync(string path);
+		Task RefreshPinnedFoldersAsync();
+		Task RemoveFolderAsync(string path);
+		Task<IEnumerable<string>> GetFoldersAsync();
+	}
 }
-
